// src/socketio/type.ts
export enum EEventMessage {
  CREATE_MESSAGE = "create_message",
  RECENT_MESSAGES = "recent_messages",
  CONVERSATION_MESSAGES = "conversation_messages",
  JOIN_ROOM = "join_room",
  NEW_MESSAGE = "new_message",
}
export interface ConversationRq {
  userId: string;
  otherUserId: string;
  limit: number;
  page: number;
}